refactor: move SshCopyIdWindow to module
